QVAC-17990 Share diffusion backend loading
