+++ packages/diffusion-cpp/addon/src/utils/BackendLoader.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <string>
+
+namespace qvac_lib_inference_addon_sd {
+
+void loadBackendModulesOnce(const std::string& backendsDir);
+
+} // namespace qvac_lib_inference_addon_sd
